requirements.txt
python-dotenv
psycopg2-binary
cryptography
discord.py
openai
